requirements.txt
matplotlib>=3.4.0
numpy>=1.20.3
pandas>=1.3.0
PyYAML
scikit_learn
scipy>=1.7.1
seaborn>=0.11.0
tensorboard
torch>=1.13.0
typing_extensions>=4.8.0
